// types.ts
export interface Photo {
  id: string;
  url: string;
  isMain: boolean;
}

export type BleachingLevel = 'None' | 'Mild' | 'Medium' | 'Strong';

export interface HealthReport {
  id: string;
  date: string;
  healthPercentage: number; // 0-100
  notes: string;
  bleaching: BleachingLevel;
}

export interface GrowthReport {
  id: string;
  date: string;
  surfaceAreaM2: number; // in square meters
  volumeM3: number; // in cubic meters
}

export interface Site {
  id: string;
  name: string;
  photoUrl: string;
  isArchived?: boolean;
}

export interface CollectionZone {
  id: string;
  name: string;
  siteId: string;
  latitude?: number;
  longitude?: number;
  isArchived?: boolean;
}

export interface Anchor {
  id:string;
  name: string;
  siteId: string;
  latitude?: number;
  longitude?: number;
  isDeepwater?: boolean;
  depth?: number;
  isArchived?: boolean;
}

export interface Tree {
  id: string;
  number: number;
  anchorId: string;
  currentDepth: number; // in meters, e.g., 8, 10, 12...
  normalDepth: number; // in meters
  lastMovedDate?: string; // ISO string
  isArchived?: boolean;
}

export interface Float {
  id: string;
  name: string;
  treeId: string;
}
// ...
export interface ScheduleItem {
  tree: Tree;
  fromDepth: number;
  toDepth: number;
}
// ...
export interface CoralBranch {
  id:string;
  fragmentId: string;
  genus: string;
  species: string;
  dateAdded: string; // ISO string format
  anchor: string;
  tree: number;
  face: 1 | 2 | 3 | 4;
  position: number;
  collectionZone: string;
  site: string;
  photos: Photo[];
  healthReports: HealthReport[];
  growthReports: GrowthReport[];
  isHeatTolerant?: boolean;
  isArchived?: boolean;
}